+++ src/devices/frameTransformStorage/FrameTransformStorage.h
@@ -13,7 +13,7 @@
 #include <yarp/sig/Vector.h>
 
 #include <yarp/dev/IFrameTransformStorage.h>
-#include <FrameTransformContainer.h>
+#include <yarp/dev/FrameTransformContainer.h>
 #include <yarp/dev/PolyDriver.h>
 #include <yarp/dev/WrapperSingle.h>
 #include <mutex>
